Removed index.js manifest, replaced it with the gulp-compiled pack.js (now index.js).
